+++ src/parse/mod.rs
@@ -2,6 +2,7 @@ mod json;
 mod jsonl;
 mod syntax;
 mod toml;
+mod xml;
 mod yaml;
 
 pub use syntax::SyntaxToken;
@@ -15,9 +16,12 @@ pub enum ContentType {
     Json,
     Yaml,
     Toml,
+    Xml,
 
-    /// Useful for some logs file: https://jsonlines.org/
+    /// Useful for some logs file: <https://jsonlines.org/>
     Jsonl,
+    // TODO: check out json-seq as specified in RFC7464 (https://datatracker.ietf.org/doc/html/rfc7464)?
+    // basically jsonl but every json object is prefixed with 0x1e
 }
 
 pub trait Parser {
@@ -25,17 +29,16 @@ pub trait Parser {
 
     fn parse(&self, data: &str) -> Result<Value>;
 
-    fn to_string(&self, value: &Value) -> String;
-
-    fn syntax_highlight(&self, value: &Value) -> Vec<SyntaxToken>;
+    fn syntax_highlight(&self, name: &str, value: &Value) -> Vec<SyntaxToken>;
 }
 
 impl ContentType {
-    pub fn new_parser(&self) -> Box<dyn Parser> {
+    pub fn new_parser(self) -> Box<dyn Parser> {
         match self {
             Self::Json => Box::new(json::JsonParser {}),
             Self::Yaml => Box::new(yaml::YamlParser {}),
             Self::Toml => Box::new(toml::TomlParser {}),
+            Self::Xml => Box::new(xml::XmlParser {}),
             Self::Jsonl => Box::new(jsonl::JsonlParser {}),
         }
     }