+++ src/parse/toml.rs
@@ -18,17 +18,8 @@ impl Parser for TomlParser {
         Ok(toml_value_to_json(toml_value))
     }
 
-    fn to_string(&self, value: &Value) -> String {
-        if let Value::Array(arr) = value {
-            // TOML does not support direct array, fallback to JSON schama.
-            return serde_json::to_string_pretty(arr).expect("serialize JSON");
-        }
-        toml::to_string_pretty(value).expect("serialize TOML")
-    }
-
-    fn syntax_highlight(&self, value: &Value) -> Vec<SyntaxToken> {
+    fn syntax_highlight(&self, _name: &str, value: &Value) -> Vec<SyntaxToken> {
         if let Value::Array(_) = value {
-            // TOML does not support direct array, fallback to JSON schama.
             return json_highlight(value, 0, false);
         }
         let mut tokens = highlight(value, None, false, false);
@@ -72,13 +63,15 @@ fn toml_value_to_json(toml_value: TomlValue) -> Value {
 
 fn highlight(
     value: &Value,
-    section: Option<String>,
+    section: Option<&str>,
     from_arr: bool,
     arr_complex: bool,
 ) -> Vec<SyntaxToken> {
     let mut tokens = Vec::new();
+    let section = section.as_ref();
+
     if from_arr {
-        if let Some(section) = section.as_ref() {
+        if let Some(section) = section {
             tokens.push(SyntaxToken::Break);
             tokens.push(SyntaxToken::Section(format!("[[{section}]]")));
             tokens.push(SyntaxToken::Break);
@@ -103,7 +96,7 @@ fn highlight(
                     }
                     tokens.push(SyntaxToken::String(String::from("'''")));
                 }
-                _ => tokens.push(SyntaxToken::String(format!("{s:?}"))),
+                StringValue::MultiLines(_) => tokens.push(SyntaxToken::String(format!("{s:?}"))),
             }
         }
         Value::Number(num) => tokens.push(SyntaxToken::Number(num.to_string())),
@@ -114,7 +107,7 @@ fn highlight(
 
         Value::Object(obj) => {
             if !from_arr {
-                if let Some(section) = section.as_ref() {
+                if let Some(section) = section {
                     tokens.push(SyntaxToken::Break);
                     tokens.push(SyntaxToken::Section(format!("[{section}]")));
                     tokens.push(SyntaxToken::Break);
@@ -144,7 +137,7 @@ fn highlight(
                     None => field,
                 };
 
-                let value_tokens = highlight(value, Some(child_section), false, true);
+                let value_tokens = highlight(value, Some(&child_section), false, true);
                 tokens.extend(value_tokens);
             }
 
@@ -163,7 +156,7 @@ fn highlight(
             debug_assert!(section.is_some());
 
             for value in arr {
-                let value_tokens = highlight(value, section.clone(), true, false);
+                let value_tokens = highlight(value, section.copied(), true, false);
                 tokens.extend(value_tokens);
             }
 
@@ -218,7 +211,7 @@ mod test {
         let parser = TomlParser {};
         for (raw, expect) in test_cases {
             let value = parser.parse(raw).unwrap();
-            let tokens = parser.syntax_highlight(&value);
+            let tokens = parser.syntax_highlight("", &value);
             let result = SyntaxToken::pure_text(&tokens);
             assert_eq!(result, expect);